+++ packages/eslint-plugin-template/src/rules/attributes-order.ts
@@ -1,12 +1,16 @@
 import type {
-  TmplAstElement,
   TmplAstBoundAttribute,
   TmplAstBoundEvent,
+  TmplAstElement,
+  TmplAstNode,
   TmplAstReference,
   TmplAstTextAttribute,
-  TmplAstNode,
+  TmplAstVariable,
 } from '@angular-eslint/bundled-angular-compiler';
-import { TmplAstTemplate } from '@angular-eslint/bundled-angular-compiler';
+import {
+  ParseSourceSpan,
+  TmplAstTemplate,
+} from '@angular-eslint/bundled-angular-compiler';
 import { getTemplateParserServices } from '@angular-eslint/utils';
 import type { TSESTree } from '@typescript-eslint/utils';
 import { createESLintRule } from '../utils/create-eslint-rule';
@@ -131,7 +135,11 @@ export default createESLintRule<Options, MessageIds>({
     }
 
     return {
-      Element$1(node: ExtendedTmplAstElement) {
+      ['Element$1, Template'](node: ExtendedTmplAstElement | TmplAstTemplate) {
+        if (isImplicitTemplate(node)) {
+          return;
+        }
+
         const { attributes, inputs, outputs, references } = node;
         const { extractedBananaBoxes, extractedInputs, extractedOutputs } =
           normalizeInputsOutputs(
@@ -227,7 +235,10 @@ function byOrder(order: readonly OrderType[], alphabetical: boolean) {
       getOrderIndex(one, order) - getOrderIndex(other, order);
 
     if (alphabetical && orderComparison === 0) {
-      return one.name > other.name ? 1 : -1;
+      return (one.keySpan?.details ?? one.name) >
+        (other.keySpan?.details ?? other.name)
+        ? 1
+        : -1;
     }
 
     return orderComparison;
@@ -238,7 +249,9 @@ function getOrderIndex(attr: ExtendedAttribute, order: readonly OrderType[]) {
   return order.indexOf(attr.orderType);
 }
 
-function toAttributeBindingOrderType(attribute: TmplAstTextAttribute) {
+function toAttributeBindingOrderType(
+  attribute: TmplAstTextAttribute | TmplAstVariable,
+) {
   return {
     ...attribute,
     orderType: OrderType.AttributeBinding,
@@ -277,12 +290,61 @@ function toTemplateReferenceVariableOrderType(reference: TmplAstReference) {
   } as ExtendedTmplAstReference;
 }
 
+function isImplicitTemplate(
+  node: TmplAstNode,
+): node is TmplAstTemplate & { tagName: null } {
+  return isTmplAstTemplate(node) && node.tagName !== 'ng-template';
+}
+
 function extractTemplateAttrs(
-  node: ExtendedTmplAstElement,
+  node: ExtendedTmplAstElement | TmplAstTemplate,
 ): (ExtendedTmplAstBoundAttribute | ExtendedTmplAstTextAttribute)[] {
-  return isTmplAstTemplate(node.parent)
-    ? node.parent.templateAttrs.map(toStructuralDirectiveOrderType)
-    : [];
+  if (isTmplAstTemplate(node)) {
+    return node.templateAttrs.map(toStructuralDirectiveOrderType).concat(
+      node.variables.map((x) => {
+        return {
+          ...toAttributeBindingOrderType(x),
+          // `let-` is excluded from the keySpan and name - add it back in
+          keySpan: new ParseSourceSpan(
+            x.keySpan.start.moveBy(-4),
+            x.keySpan.end,
+          ),
+          name: 'let-' + x.name,
+        } as ExtendedTmplAstTextAttribute;
+      }),
+    );
+  }
+
+  if (!isImplicitTemplate(node.parent)) {
+    return [];
+  }
+
+  /*
+   * There may be multiple "attributes" for a structural directive even though
+   * there is only a single HTML attribute:
+   * e.g. `<ng-container *ngFor="let foo of bar"></ng-container>`
+   * will parsed as two attributes (`ngFor` and `ngForOf`)
+   */
+
+  const attrs = node.parent.templateAttrs.map(toStructuralDirectiveOrderType);
+
+  // Pick up on any subsequent `let` bindings, e.g. `index as i`
+  let sourceEnd = attrs[attrs.length - 1].sourceSpan.end;
+  node.parent.variables.forEach((v) => {
+    if (
+      v.sourceSpan.start.offset <= sourceEnd.offset &&
+      sourceEnd.offset < v.sourceSpan.end.offset
+    ) {
+      sourceEnd = v.sourceSpan.end;
+    }
+  });
+
+  return [
+    {
+      ...attrs[0],
+      sourceSpan: new ParseSourceSpan(attrs[0].sourceSpan.start, sourceEnd),
+    } as ExtendedTmplAstBoundAttribute | ExtendedTmplAstTextAttribute,
+  ];
 }
 
 function normalizeInputsOutputs(
@@ -333,19 +395,20 @@ function isOnSameLocation(
 }
 
 function getMessageName(expected: ExtendedAttribute): string {
+  const fullName = expected.keySpan?.details ?? expected.name;
   switch (expected.orderType) {
     case OrderType.StructuralDirective:
-      return `*${expected.name}`;
+      return `*${fullName}`;
     case OrderType.TemplateReferenceVariable:
-      return `#${expected.name}`;
+      return `#${fullName}`;
     case OrderType.InputBinding:
-      return `[${expected.name}]`;
+      return `[${fullName}]`;
     case OrderType.OutputBinding:
-      return `(${expected.name})`;
+      return `(${fullName})`;
     case OrderType.TwoWayBinding:
-      return `[(${expected.name})]`;
+      return `[(${fullName})]`;
     default:
-      return expected.name;
+      return fullName;
   }
 }