+++ packages/eslint-plugin-template/tests/rules/attributes-order/cases.ts
@@ -1,6 +1,8 @@
 import { convertAnnotatedSourceToFailureCase } from '@angular-eslint/utils';
-import type { MessageIds } from '../../../src/rules/attributes-order';
-import { OrderType } from '../../../src/rules/attributes-order';
+import {
+  OrderType,
+  type MessageIds,
+} from '../../../src/rules/attributes-order';
 
 const messageId: MessageIds = 'attributesOrder';
 
@@ -11,9 +13,34 @@ export const valid = [
   '<input *ngIf="flag" required>',
   `<input [(ngModel)]="model">`,
   '<input [(ngModel)]="model" (ngModelChange)="onChange($event)">',
+  '<ng-template></ng-template>',
+  '<ng-template #Template><div></div></ng-template>',
+  '<ng-template [ngIf]="condition" [ngIfThen]="If" [ngIfElse]="Else"><div></div></ng-template>',
+  '<ng-template #Template let-value><div></div></ng-template>',
 ];
 
 export const invalid = [
+  convertAnnotatedSourceToFailureCase({
+    messageId,
+    description: 'should work for simple attributes',
+    annotatedSource: `
+      <li><input type="text" id="input"></li>
+                 ~~~~~~~~~~~~~~~~~~~~~~
+    `,
+    options: [
+      {
+        alphabetical: true,
+      },
+    ],
+    data: {
+      expected: '`id`, `type`',
+      actual: '`type`, `id`',
+    },
+    annotatedOutput: `
+      <li><input id="input" type="text"></li>
+                 
+    `,
+  }),
   convertAnnotatedSourceToFailureCase({
     messageId,
     description: 'should fail if structural directive is not first',
@@ -193,6 +220,122 @@ export const invalid = [
         
         (output)="handleOutput($event)">
         
+    `,
+  }),
+  convertAnnotatedSourceToFailureCase({
+    messageId,
+    description: 'should work for ngFor',
+    annotatedSource: `
+      <ng-container (click)="bar = []" id="issue" *ngFor="let foo of bar"></ng-container>
+                    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    `,
+    data: {
+      expected: '`*ngFor`, `id`, `(click)`',
+      actual: '`(click)`, `id`, `*ngFor`',
+    },
+    annotatedOutput: `
+      <ng-container *ngFor="let foo of bar" id="issue" (click)="bar = []"></ng-container>
+                    
+    `,
+  }),
+  convertAnnotatedSourceToFailureCase({
+    messageId,
+    description: 'should work for ngFor with let',
+    annotatedSource: `
+      <ng-container (click)="bar = []" id="issue" *ngFor="let foo of bar; index as i; first as isFirst"></ng-container>
+                    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    `,
+    data: {
+      expected: '`*ngFor`, `id`, `(click)`',
+      actual: '`(click)`, `id`, `*ngFor`',
+    },
+    annotatedOutput: `
+      <ng-container *ngFor="let foo of bar; index as i; first as isFirst" id="issue" (click)="bar = []"></ng-container>
+                    
+    `,
+  }),
+  convertAnnotatedSourceToFailureCase({
+    messageId,
+    description: 'should work for ngIf as',
+    annotatedSource: `
+      <div id="id" *ngIf="bar as foo"></div>
+           ~~~~~~~~~~~~~~~~~~~~~~~~~~
+    `,
+    data: {
+      expected: '`*ngIf`, `id`',
+      actual: '`id`, `*ngIf`',
+    },
+    annotatedOutput: `
+      <div *ngIf="bar as foo" id="id"></div>
+           
+    `,
+  }),
+  convertAnnotatedSourceToFailureCase({
+    messageId,
+    description: 'should work for ngIfThenElse',
+    annotatedSource: `
+      <div id="id" *ngIf="condition then foo else bar"></div>
+           ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    `,
+    data: {
+      expected: '`*ngIf`, `id`',
+      actual: '`id`, `*ngIf`',
+    },
+    annotatedOutput: `
+      <div *ngIf="condition then foo else bar" id="id"></div>
+           
+    `,
+  }),
+  convertAnnotatedSourceToFailureCase({
+    messageId,
+    description: 'should work for dotted attributes',
+    annotatedSource: `
+      <div [disabled]="disabled" [class.disabled]="disabled"></div>
+           ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    `,
+    options: [{ alphabetical: true }],
+    data: {
+      expected: '`[class.disabled]`, `[disabled]`',
+      actual: '`[disabled]`, `[class.disabled]`',
+    },
+    annotatedOutput: `
+      <div [class.disabled]="disabled" [disabled]="disabled"></div>
+           
+    `,
+  }),
+  convertAnnotatedSourceToFailureCase({
+    messageId,
+    description: 'should work with ng-template',
+    annotatedSource: `
+      <ng-template let-value #Template></ng-template>
+                   ~~~~~~~~~~~~~~~~~~~
+    `,
+    options: [{ alphabetical: true }],
+    data: {
+      expected: '`#Template`, `let-value`',
+      actual: '`let-value`, `#Template`',
+    },
+    annotatedOutput: `
+      <ng-template #Template let-value></ng-template>
+                   
+    `,
+  }),
+  convertAnnotatedSourceToFailureCase({
+    messageId,
+    description:
+      'should work with ng-template with multiple variable assignments',
+    annotatedSource: `
+      <ng-template let-value="something" let-anotherValue="else" #Template></ng-template>
+                   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    `,
+    options: [{ alphabetical: true }],
+    data: {
+      expected: '`#Template`, `let-anotherValue`, `let-value`',
+      actual: '`let-value`, `let-anotherValue`, `#Template`',
+    },
+    annotatedOutput: `
+      <ng-template #Template let-anotherValue="else" let-value="something"></ng-template>
+                   
     `,
   }),
 ];