+++ packages/eslint-plugin-template/tests/rules/use-track-by-function/cases.ts
@@ -52,6 +52,30 @@ export const valid = [
         trackBy : trackByFn
       ">
     `,
+  {
+    options: [{ alias: ['ngForTrackByProperty'] }],
+    code: `
+        <div *ngFor="
+          let item of [1, 2, 3];
+          let i = index;
+          trackByProperty: 'id'
+        ">
+      `,
+  },
+  {
+    options: [{ alias: ['ngForTrackById'] }],
+    code: `
+      <div *ngFor="let photo of photos; trackById"></div>
+    `,
+  },
+  {
+    options: [{ alias: ['ngForTrackByProperty'] }],
+    code: `
+      <ng-template ngFor let-item [ngForOf]="[1, 2, 3]" let-i="index" [ngForTrackByProperty]="trackByFn">
+        {{ item }}
+      </ng-template>
+    `,
+  },
 ];
 
 export const invalid = [
@@ -68,7 +92,33 @@ export const invalid = [
     messageId,
   }),
   convertAnnotatedSourceToFailureCase({
+    description:
+      'should fail when trackBy function (`ngForTrackByProperty` alias) is not present',
+    options: [{ alias: ['ngForTrackByProperty'] }],
+    annotatedSource: `
+        <ul>
+          <li *ngFor="let item of [1, 2, 3];">
+              ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+            {{ item }}
+          </li>
+        </ul>
+      `,
+    messageId,
+  }),
+  convertAnnotatedSourceToFailureCase({
     description: 'should fail when [ngForTrackBy] is missing in ng-template',
+    annotatedSource: `
+        <ng-template ngFor let-item [ngForOf]="[1, 2, 3]" let-i="index">
+                                    ~~~~~~~~~~~~~~~~~~~~~
+          {{ item }}
+        </ng-template>
+      `,
+    messageId,
+  }),
+  convertAnnotatedSourceToFailureCase({
+    description:
+      'should fail when [ngForTrackByProperty] is missing in ng-template',
+    options: [{ alias: ['ngForTrackByProperty'] }],
     annotatedSource: `
         <ng-template ngFor let-item [ngForOf]="[1, 2, 3]" let-i="index">
                                     ~~~~~~~~~~~~~~~~~~~~~
@@ -116,5 +166,25 @@ export const invalid = [
         messageId,
       },
     ],
+  }),
+  convertAnnotatedSourceToFailureCase({
+    description:
+      'should fail when there are three ngFor and for the third, trackBy function (`ngForTrackByProperty` alias) is not present',
+    options: [{ alias: ['ngForTrackByProperty'] }],
+    annotatedSource: `
+        <div *ngFor="let item of [1, 2, 3]; trackBy: trackByFn">
+          {{ item }}
+        </div>
+        <div *ngFor="let item of [1, 2, 3]; trackByProperty: trackByFn">
+          {{ item }}
+        </div>
+        <ul>
+          <li *ngFor="let item of [1, 2, 3];">
+              ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+            {{ item }}
+          </li>
+        </ul>
+      `,
+    messageId,
   }),
 ];